kk3

+++ libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: amzahir <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/07 18:47:05 by amzahir           #+#    #+#             */
-/*   Updated: 2024/12/07 18:52:30 by amzahir          ###   ########.fr       */
+/*   Updated: 2024/12/08 17:25:37 by amzahir          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include <unistd.h>
 # include <stdarg.h>
 
-void	ft_putnbr_base(int nb, char *base);
+void	ft_putnbr_base(long nb, char *base);
 int		ft_printf(const char *fmt, ...);
 
 #endif
